Fix budget was incorrect for month >= September 2019 due to month count resetting back to 0.

+++ android-app/app/src/main/java/com/georgeflug/budget/view/budget/MonthCategoryRollup.kt
@@ -4,7 +4,7 @@ import com.georgeflug.budget.model.Budget
 import com.georgeflug.budget.model.TransactionSplit
 import java.math.BigDecimal
 
-class MonthCategoryRollup(val budget: Budget, carryoverMonths: Int) {
+class MonthCategoryRollup(val budget: Budget, carryoverMonths: Long) {
     val title = budget.title
     val iconId = budget.iconId
     var total: BigDecimal = BigDecimal.ZERO
